@@ -4,7 +4,7 @@ version: 0.2.0+2
 homepage: https://github.com/ferraridamiano/units_converter
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0-259.9.beta <3.0.0'
 
 dev_dependencies:
   pedantic: ^1.9.0
