@@ -1,7 +1,7 @@
 class Unit {
-  double value;
-  String stringValue;
+  double? value;
+  String? stringValue;
   var name;
-  String symbol;
+  String? symbol;
   Unit(this.name, {this.value, this.stringValue, this.symbol});
 }
