@@ -85,12 +85,12 @@ class Area extends Property<AREA, double> {
 
   ///Converts a unit with a specific name (e.g. AREA.hectares) and value to all other units
   @override
-  void convert(AREA name, double value) {
+  void convert(AREA name, double? value) {
     super.convert(name, value);
     if (value == null) return;
     for (var i = 0; i < AREA.values.length; i++) {
-      unitList[i].value = unit_conversion.getByName(AREA.values.elementAt(i)).value;
-      unitList[i].stringValue = mantissaCorrection(unitList[i].value, significantFigures, removeTrailingZeros);
+      unitList[i].value = unit_conversion.getByName(AREA.values.elementAt(i))?.value;
+      unitList[i].stringValue = mantissaCorrection(unitList[i].value!, significantFigures, removeTrailingZeros);
     }
   }
 
