@@ -55,12 +55,12 @@ class Force extends Property<FORCE, double> {
 
   ///Converts a unit with a specific name (e.g. FORCE.newton) and value to all other units
   @override
-  void convert(FORCE name, double value) {
+  void convert(FORCE name, double? value) {
     super.convert(name, value);
     if (value == null) return;
     for (var i = 0; i < FORCE.values.length; i++) {
-      unitList[i].value = unit_conversion.getByName(FORCE.values.elementAt(i)).value;
-      unitList[i].stringValue = mantissaCorrection(unitList[i].value, significantFigures, removeTrailingZeros);
+      unitList[i].value = unit_conversion.getByName(FORCE.values.elementAt(i))?.value;
+      unitList[i].stringValue = mantissaCorrection(unitList[i].value!, significantFigures, removeTrailingZeros);
     }
   }
 
