@@ -57,12 +57,12 @@ class Pressure extends Property<PRESSURE, double> {
 
   ///Converts a unit with a specific name (e.g. PRESSURE.psi) and value to all other units
   @override
-  void convert(PRESSURE name, double value) {
+  void convert(PRESSURE name, double? value) {
     super.convert(name, value);
     if (value == null) return;
     for (var i = 0; i < PRESSURE.values.length; i++) {
-      unitList[i].value = unit_conversion.getByName(PRESSURE.values.elementAt(i)).value;
-      unitList[i].stringValue = mantissaCorrection(unitList[i].value, significantFigures, removeTrailingZeros);
+      unitList[i].value = unit_conversion.getByName(PRESSURE.values.elementAt(i))?.value;
+      unitList[i].stringValue = mantissaCorrection(unitList[i].value!, significantFigures, removeTrailingZeros);
     }
   }
 
